Move shared fixtures to a higher conftest level
